カート処理修正

+++ backend/src/main/java/ec/training/mapper/CartsEntityMapper.java
@@ -38,6 +38,5 @@ public interface CartsEntityMapper {
      * @param userId
      * @return ユーザーに紐づくカート
      */
-    Optional<CartsEntity> selectByUserId(Integer userId);
-
+    Optional<CartsEntity> selectByUserId(Long userId);
 }
